# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parlours Booking API",
        "description": "Swagger API Documentation in Parlours Booking",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/userMobileRegister": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registeration User Mobile No",
                "description": "Returns user OTP",
                "operationId": "userRegisterMobileNo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usertype",
                                    "phone_no"
                                ],
                                "properties": {
                                    "usertype": {
                                        "type": "number",
                                        "example": "2(Customer)/1(Vendor)"
                                    },
                                    "phone_no": {
                                        "type": "number",
                                        "example": "0711234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Mobile No Registered Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "OTP": {
                                            "type": "string",
                                            "example": "Send to mobile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/userMobileVerification": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User mobile no Verification",
                "description": "Returns user token",
                "operationId": "verifyVerificationCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "verification_code"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "example": "2"
                                    },
                                    "verification_code": {
                                        "type": "number",
                                        "example": "OTG478"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Verified Successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "generated_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/userRegistration": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Registration",
                "description": "Register the user(Vendor/Customer) basic details",
                "operationId": "userRegistration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usertype",
                                    "first_name",
                                    "last_name",
                                    "address",
                                    "city",
                                    "dob",
                                    "gender",
                                    "phone_no",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-08-15"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male"
                                    },
                                    "phone_no": {
                                        "type": "string",
                                        "example": "0711234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "vendor_type": {
                                        "description": "Required only if usertype is 1 (Vendor)",
                                        "type": "string",
                                        "example": "Grocery Store",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Registered Successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "generated_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/userSetNewPassword": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User New Password Setup",
                "description": "Setup New password for registered user",
                "operationId": "userSetNewPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User New Password Set Successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "generated_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/userLogin": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login User",
                "description": "Returns user token",
                "operationId": "userLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_no",
                                    "password"
                                ],
                                "properties": {
                                    "phone_no": {
                                        "type": "number",
                                        "example": "0711234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "generated_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}